Refactoring interface
The old line number in each task is removed, now is not stored the line
number it is calculated each time on the interface class. That
refactoring will be necessary to make the title editor multi-line.

That could make the program slower, maybe one day will be necessary to
use a kind of cache of the task lines on screen so they are not
calculated each time.